requirements.txt
# Backend dependencies
flask==2.0.1
flask-cors==3.0.10
requests==2.26.0
python-dotenv==0.19.0
biopython==1.79
# Fix compatibility issues
werkzeug==2.0.1
numpy==1.20.3
pandas==1.3.3
# For parsing and processing XML/HTML
beautifulsoup4==4.10.0
lxml==4.6.3
# LangChain for PubMed integration
langchain>=0.0.267
langchain-community>=0.0.10